Add brand-new Section retrieving function
These changes are necessary to simplify the work of the "Dovecote" section. Now the whole tree will be assembled at once and sent to the frontend, rather than being assembled gradually, as needed. The subject area allows  to apply this approach. There can hardly be so many sections and pigeons that it will slow down the application.