Make select and text input controlled
Converts the form and its components into a controlled form. Two types of input fields are being refactored here, and two will be refactored in future commits.
The decision is partly due to the transition to the Material library, and partly due to the desire to implement behavior that is easier to implement using controlled components.