Add functions for creating hierarchical view
In addition to converting hierarchical objects into a flat array with modified names, such names are required separately - without converting one object to another. For this, functions are added that use similar mechanics and add a corresponding name to each hierarchical object. All this serves to form a hierarchical representation in the drop-down lists of various input fields.